Treat fanout (client) subscriptions as a no-op
As documented in [issue #5][1], clients with subscriptions defined encounter
exceptions when they attempt to process check results as check requests.

With Sensu's forthcoming [native silencing implementation][2], each client
definition will include a subscription for "client:$CLIENT_NAME". For the
purposes of this silencing implementation, it's not important that the client
consume check requests from this subscription, but it is important that the
subscription is present in the client definition sent via keepalive messages.

When Sensu processes call `subscribe()` on the transport object, they pass a
queue type. Sensu server processes subscribe with type `:direct` while
Sensu client processes use type `:fanout`. This distinction makes it possible
for us to treat client subscriptions for this transport as a no-op.

[1]: https://github.com/SimpleFinance/sensu-transport-snssqs/issues/5
[2]: https://github.com/sensu/sensu/issues/1328